fix(cli): wire service add (bare) to the wizard, remove nyxid wizard alias
Per docs/CLI_WIZARD_V2.md §3.1 + §6, the primary entry point for the
M3 wizard is 'nyxid service add' (the existing command), not a new
'nyxid wizard' top-level subcommand. The previous commit added the
alias without ever touching service add, so users running the natural
command hit the old non-interactive path.

Fix:

- Remove Commands::Wizard + WizardArgs from cli.rs
- Remove run_wizard dispatch from main.rs
- Move the wizard entry (resolve auth + run flow + print summary) into
  wizard::run_ai_key_wizard so there is one shared function
- Add a bare-invocation check at the top of ServiceCommands::Add:
    slug.is_none() && !custom && !oauth && !device_code
    && via_node.is_none() && endpoint_url.is_none() && label.is_none()
    && auth_method.is_none() && auth_key_name.is_none()
    && credential.is_none() && credential_env.is_none()
    && scopes.is_empty() && org.is_none() && openapi_spec_url.is_none()
    && output == Table && stdout is a TTY
  → dispatch to wizard::run_ai_key_wizard

Behavior matrix:

- nyxid service add (TTY)                    → wizard opens
- nyxid service add (non-TTY / piped)        → existing 'provide slug' error
- nyxid service add --output json            → existing scripted path
- nyxid service add llm-openai --credential… → existing scripted path
- nyxid service add --oauth                  → existing OAuth path
- nyxid service add --custom                 → existing custom-endpoint path

Scripted and CI invocations are byte-identical to before. Only truly
bare interactive invocations are affected.
